moved our season stats methods into its own module
